Stop supporting PHP 8.0
