+++ .gitignore
@@ -1,4 +1,5 @@
 composer.lock
+.php-version
 /vendor
 /.phpunit.cache
 .phpunit.result.cache
