+++ PHP_VERSIONS
@@ -1,4 +1,3 @@
 8.3
 8.2
 8.1
-8.0
