+++ bin/TestAndLint.sh
@@ -55,7 +55,7 @@ test_and_lint() {
     echo "-----------------------------------------------------------"
     echo "[PHP $1][phpunit]"
     ./vendor/bin/phpunit ./tests/ \
-                         --color auto
+                         --color=auto
     echo "-----------------------------------------------------------"
 }
