+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "macocci7/php-frequency-table",
-    "version": "1.2.0",
+    "version": "1.3.0",
     "description": "it's easy to use for creating frequency distribution tables.",
     "type": "library",
     "license": "MIT",
@@ -20,12 +20,12 @@
         "squizlabs/php_codesniffer": "^3.7",
         "phpstan/phpstan": "^1.10",
         "phpmd/phpmd": "^2.15",
-        "phpunit/phpunit": "9.6",
-        "phploc/phploc": "^7.0",
+        "phpunit/phpunit": "^10.5",
         "php-parallel-lint/php-parallel-lint": "1.3",
         "macocci7/php-combination": "^1.0"
     },
     "require": {
+        "php": ">=8.1",
         "nette/neon": "^3.4"
     }
 }
