+++ tests/Helper/ConfigTest.php
@@ -4,15 +4,12 @@ declare(strict_types=1);
 
 namespace Macocci7\PhpFrequencyTable\Helper;
 
-require('vendor/autoload.php');
-
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Macocci7\PhpFrequencyTable\Helper\Config;
 use Nette\Neon\Neon;
 
 /**
- * @SuppressWarnings(PHPMD.TooManyMethods)
- * @SuppressWarnings(PHPMD.TooManyPublicMethods)
  * @SuppressWarnings(PHPMD.CamelCaseMethodName)
  */
 final class ConfigTest extends TestCase
@@ -61,9 +58,7 @@ final class ConfigTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider provide_className_can_return_class_name_correctly
-     */
+    #[DataProvider('provide_className_can_return_class_name_correctly')]
     public function test_className_can_return_class_name_correctly(string $class, string $expect): void
     {
         $this->assertSame($expect, Config::className($class));
@@ -95,9 +90,7 @@ final class ConfigTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider provide_support_object_like_keys_correctly
-     */
+    #[DataProvider('provide_support_object_like_keys_correctly')]
     public function get_can_support_object_like_keys_correctly(string $key, array|null $expect): void
     {
         $this->assertSame($expect, Config::get($key));
